init commit

+++ judge/views/problem_data.py
@@ -9,7 +9,8 @@ from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.exceptions import ValidationError
-from django.forms import BaseModelFormSet, HiddenInput, ModelForm, NumberInput, Select, formset_factory
+from django.forms import BaseModelFormSet, HiddenInput, ModelForm, NumberInput, Select, Textarea, formset_factory
+from django.forms.models import inlineformset_factory
 from django.http import Http404, HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
@@ -19,7 +20,9 @@ from django.utils.translation import gettext as _
 from django.views.generic import DetailView
 
 from judge.highlight_code import highlight_code
-from judge.models import Problem, ProblemData, ProblemTestCase, Submission, problem_data_storage
+from django_ace import AceWidget
+
+from judge.models import Problem, ProblemData, ProblemHarness, ProblemTestCase, Submission, problem_data_storage
 from judge.utils.problem_data import ProblemDataCompiler
 from judge.utils.unicode import utf8text
 from judge.utils.views import TitleMixin, add_file_response
@@ -102,6 +105,23 @@ class ProblemCaseFormSet(formset_factory(ProblemCaseForm, formset=BaseModelFormS
         return form
 
 
+class ProblemHarnessForm(ModelForm):
+    class Meta:
+        model = ProblemHarness
+        fields = ['language', 'entry_point', 'skip_precompile', 'harness_code']
+        widgets = {
+            'harness_code': AceWidget(mode='java', theme='chrome', width='100%', height='300px'),
+        }
+
+
+ProblemHarnessFormSet = inlineformset_factory(
+    Problem, ProblemHarness,
+    form=ProblemHarnessForm,
+    extra=1,
+    can_delete=True,
+)
+
+
 class ProblemManagerMixin(LoginRequiredMixin, ProblemMixin, DetailView):
     def get_object(self, queryset=None):
         problem = super(ProblemManagerMixin, self).get_object(queryset)
@@ -171,6 +191,13 @@ class ProblemDataView(TitleMixin, ProblemManagerMixin):
         return ProblemCaseFormSet(data=self.request.POST if post else None, prefix='cases', valid_files=files,
                                   queryset=ProblemTestCase.objects.filter(dataset_id=self.object.pk).order_by('order'))
 
+    def get_harness_formset(self, post=False):
+        return ProblemHarnessFormSet(
+            data=self.request.POST if post else None,
+            instance=self.object,
+            prefix='harnesses',
+        )
+
     def get_valid_files(self, data, post=False) -> List[str]:
         try:
             if post and 'problem-data-zipfile-clear' in self.request.POST:
@@ -197,6 +224,9 @@ class ProblemDataView(TitleMixin, ProblemManagerMixin):
 
         context['cases_formset'] = self.get_case_formset(valid_files)
         context['all_case_forms'] = chain(context['cases_formset'], [context['cases_formset'].empty_form])
+
+        if 'harness_formset' not in context:
+            context['harness_formset'] = self.get_harness_formset()
         return context
 
     def post(self, request, *args, **kwargs):
@@ -210,17 +240,24 @@ class ProblemDataView(TitleMixin, ProblemManagerMixin):
             data_form.zip_valid = False
 
         cases_formset = self.get_case_formset(valid_files, post=True)
-        if data_form.is_valid() and cases_formset.is_valid():
+        harness_formset = self.get_harness_formset(post=True)
+
+        if data_form.is_valid() and cases_formset.is_valid() and harness_formset.is_valid():
             data = data_form.save()
             for case in cases_formset.save(commit=False):
                 case.dataset_id = problem.id
                 case.save()
             for case in cases_formset.deleted_objects:
                 case.delete()
+
+            # Save harness formset
+            harness_formset.save()
+
             ProblemDataCompiler.generate(problem, data, problem.cases.order_by('order'), valid_files)
             return HttpResponseRedirect(request.get_full_path())
-        return self.render_to_response(self.get_context_data(data_form=data_form, cases_formset=cases_formset,
-                                                             valid_files=valid_files))
+        return self.render_to_response(self.get_context_data(
+            data_form=data_form, cases_formset=cases_formset,
+            harness_formset=harness_formset, valid_files=valid_files))
 
     put = post
 
@@ -336,7 +373,34 @@ def download_tester(request, problem):
     if not problem.view_tester:
         raise Http404()
 
-    # Check if MainTest.java exists
+    # Get available harnesses
+    harnesses = problem.harnesses.select_related('language').all()
+
+    if harnesses.exists():
+        # Get requested language from query param, or default to first harness
+        lang_key = request.GET.get('lang')
+        if lang_key:
+            harness = harnesses.filter(language__key=lang_key).first()
+        else:
+            harness = harnesses.first()
+
+        if not harness:
+            raise Http404()
+
+        # Determine filename based on language
+        if harness.language.key.startswith('JAVA'):
+            filename = '%s.java' % (harness.entry_point or 'MainTest')
+        elif harness.language.key in ('PY3', 'PYPY3', 'PY2', 'PYPY'):
+            filename = 'tester.py'
+        else:
+            filename = 'tester.txt'
+
+        # Return harness code as downloadable file
+        response = HttpResponse(harness.harness_code, content_type='application/octet-stream')
+        response['Content-Disposition'] = 'attachment; filename="%s"' % filename
+        return response
+
+    # Fall back to legacy MainTest.java if it exists
     filename = 'MainTest.java'
     path = os.path.join(problem.code, filename)
